+++ docs/whats-new.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Fix
+- removed `matplotlib._api` import to also support `matplotlib<3.4.0` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#465](https://github.com/Deltares/dfm_tools/pull/465)
+
+
 ## 0.13.1 (2023-07-12)
 
 ### Fix